more shadowing + changing size_t to int in []() because inputs are ints and internal ops are int

+++ include/igl/unique_rows.cpp
@@ -35,8 +35,8 @@ IGL_INLINE void igl::unique_rows(
     vIA[i] = i;
   }
 
-  auto index_equal = [&sortA, &num_cols](const size_t i, const size_t j) {
-    for (size_t c=0; c<num_cols; c++) {
+  auto index_equal = [&sortA, &num_cols](const int i, const int j) {
+    for (int c=0; c<num_cols; c++) {
       if (sortA(i,c) != sortA(j,c))
         return false;
     }
